Camera2: Swap AF notification order, minor bug fix.
In continuous AF modes, when an AF trigger/cancel is received in the
middle of a passive AF scan, we want to send both an AF move stop
callback and an AF done callback. Swap the order of sending to
simplify life when the app just wants to pay attention to the AF done
callback during an active AF trigger.

Bug: 6243944
Change-Id: I127d1675bc37ef828ef89c70b8cfab75f9428a2f